Add Filtering, Sorting and Pagination

+++ Walkistan.API/Models/Domain/Walk.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public double LengthInKm { get; set; }
+        public int LengthInKm { get; set; }
         public string? WalkImageUrl { get; set; }
 
         public int DifficultyId { get; set; }
