+++ Walkistan.API/Models/Dto/Walk/WalkDto.cs
@@ -8,7 +8,7 @@ namespace Walkistan.API.Models.Dto.Walk
         public int Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public double LengthInKm { get; set; }
+        public int LengthInKm { get; set; }
         public string? WalkImageUrl { get; set; }
